ci(docs): add strict test for mkdocs build on pr
Runs `mkdocs build --strict` to check for faulty docs resulting in an
abort on errors and warnings. This will not abort early for non-critical
errors, so the output should contain everything on the first run.
The project is now configured to emit warnings for link and nav related
problems so the ci will fail with the strict flag.
Testing locally is still fine without the --strict setting, so the
results can be viewed live in the browser.
The workflow uses mkdocs instead of mike, since versioned docs are not necessary
and need more ressources.