docs: switch tooling to Material for MkDocs + Mike (#446)
* docs: setup MkDocs material

* docs: remove mdbook, move src into docs dir

* docs: add mkdoxy pin mkdocs material major version

Both packages use semantic versioning, so pinnig to the current major
version should prevent some headache in case of an automatic upgrade.

* docs: add mkdocs-awesome-nav, mkdoxy examples

Page structure moved from main config to `docs/.nav.yml` file.
The `mkdocs-awesome-nav` allows us to use globs and folders for the site
structure. Without this it is impossible to modify the automatically
generated page structure without enumerating *every* page.

* docs: fix API snippets, ignore doxyfile

Renamed the Mkdoxy Project from `rayx-core` to `rayxCore` because the
dash broke snippet integration.
No longer uses the Doxyfile for settings anymore.

* ci(docs): add mkdocs and remove mdBook workflow

* docs: add logo

* ci(docs): fix typo action*s*

* docs(python): update uv lockfile

Add support for Python 3.14, which the CI defaults to now.

* docs: add favicon

* ci(docs): add edit and view src actions

* ci(docs): trigger deploy workflow on tag v*

Pushing a version tag did not result in triggering the docs deployment
action with the tag information, so the version release step was skipped.

* ci(docs): remove version trigger, use env var

Previously a push with a version tag would trigger the workflow twice.
The `--ignore-remote-status` flag was added to the mike commands to not
fail on concurrent runs of the workflow (the second run has a diverging
git history after the first one pushes).
This could still happen if there are multiple pushes before the workflow
finishes.
If a new docs release is pushed manually before the CI finishes it might
be overwritten now, whereas previously the CI would fail to push.

* ci(docs): only fetch current commit + gh-pages

* ci(docs): fix version tag retrieval

* ci(docs): fix env substitution

Can't use bash string substitution with Github context variables as
these are inserted before the step runs.

* ci(docs): add strict test for mkdocs build on pr

Runs `mkdocs build --strict` to check for faulty docs resulting in an
abort on errors and warnings. This will not abort early for non-critical
errors, so the output should contain everything on the first run.
The project is now configured to emit warnings for link and nav related
problems so the ci will fail with the strict flag.
Testing locally is still fine without the --strict setting, so the
results can be viewed live in the browser.
The workflow uses mkdocs instead of mike, since versioned docs are not necessary
and need more ressources.

* docs: add additional explanations, fix API example

* docs: fix absolute links should be relative

* docs: fix how to build anchor

* docs: fix multiple image links, missing linebreaks

* docs: fix thesis links and formatting

* docs: enable permalinks for headline anchors

These appear as little indicators next to headlines, allowing to link
directly to headline anchors.